track tests at more correct place
